refactor: List<Reservation> 관련 로직을 처리하는 도메인 생성

+++ src/main/java/roomescape/controller/reservation/dto/MyReservationResponse.java
@@ -9,9 +9,9 @@ public record MyReservationResponse(
         String themeName,
         LocalDate date,
         LocalTime time,
-        Long rank) {
-    
-    public static MyReservationResponse from(final Reservation reservation, final Long rank) {
+        int rank) {
+
+    public static MyReservationResponse from(final Reservation reservation, final int rank) {
         return new MyReservationResponse(
                 reservation.getId(),
                 reservation.getTheme().getName(),
